Add 'exec' event
On Linux and Mac, the child process is spawn using fork + exec.  There
is currently no way to tell if the child has failed to exec (for
example, when trying to execute a non-existent file, a directory, or a
file with insufficient permissions).  We just get an exit event with
code 1.

My first intention was to add an 'error' event.  However, on top of
knowing when the child's exec fails, I need to be able to tell that the
exec has succeeded, so I can report to upper layers that the process has
been successfully started.

Therefore, I instead chose to add an 'exec' event, which is emitted to
report whether the child process has successfully exec'ed or not.  If
the 'error' argument passed to the callback is undefined, it means the
exec was successful.

The implementation (using a pipe) is inspired by what libuv does, it is
explained inline.

Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>